requirements.txt
youtube-transcript-api==0.6.3
pytube
gradio==5.5.0
langchain==0.3.10
langchain_huggingface==0.1.0
langchain_community==0.3.8
langchain_groq==0.2.1
python-dotenv
qdrant_client==1.12.1
langchain_qdrant==0.2.0